refactor(crosschain): improve project structure
